+++ release_notes.md
@@ -1,3 +1,11 @@
-## Bug fixes
+# Release Notes
+
+## Microsoft.Azure.Functions.Worker.Extensions.DurableTask
+
+1.0.0 GA release of Durable Functions for .NET isolated worker. This release includes support for running Orchestrations and Activities in the isolated worker. Entities are not supported yet.
+
+## Microsoft.Azure.WebJobs.Extensions.DurableTask
+
+### Bug fixes
 
 - Fix deserialization of class-based entities to use custom serializer settings (resolves https://github.com/Azure/azure-functions-durable-extension/issues/2361)